+++ kernel/src/HAL/timer.hpp
@@ -1,0 +1,10 @@
+#ifndef _HAL_TIMER_HPP
+#define _HAL_TIMER_HPP
+
+#include <stdint.h>
+
+void HAL_TimerInit();
+
+void sleep(uint64_t ms);
+
+#endif /* _HAL_TIMER_HPP */