+++ kernel/src/Memory/PagingUtil.cpp
@@ -1,0 +1,7 @@
+#include "PagingUtil.hpp"
+
+#include <arch/x86_64/Memory/PageMapIndexer.hpp>
+
+void MapPage(void* phys_addr, void* virt_addr, uint32_t flags) {
+    x86_64_map_page(phys_addr, virt_addr, flags);
+}