+++ CHANGES.md
@@ -1,6 +1,12 @@
 # Changes
 
-## Latest Changes - 27/02/2023
+## Latest Changes - 05/03/2023
+
+- Added PIT timer with full timer interface
+- Fixed qemu run command line.
+- Moved Virtual and physical page managers out of arch/x86_64 folder
+
+## 27/02/2023
 
 - Fixed interrupt returning and ISR registering
 - Removed nasm building as it is unnecessary